VS: Out reg components can be non-sequential
In the case of Super Monkey Ball, one of the output registers has all
components set to out target 31. The old implementation was trying to
write out of bounds for the yzw components! The dest register has been
refactored to take into account that the components can come from
different, non-sequential sources.

This commit also zero fills the input vertex attributes that are not
set, which I would assume is normal hardware behaviour but needs to be
tested either way. This in itself fixes a few things.